# FeaturesProvider/openapi.yaml
openapi: 3.0.3
info:
  title: TrackMyOffer Feature Provider
  description: API for executing some AI-powered features
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /greet:
    post:
      summary: Generate greeting
      description: Generates an AI-powered greeting for a given name
      operationId: greet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "John"
      responses:
        '200':
          description: Successful greeting
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Hello, John!"
        '400':
          description: Bad request - name is missing
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Name is required"
        '503':
          description: AI service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "AI service is currently unavailable"
  /api/profile:
    post:
      summary: Create or update profile
      description: Creates a new profile or updates existing one if email already exists
      operationId: createOrUpdateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
      responses:
        '201':
          description: Profile created or updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: array
                    items:
                      type: object
                      properties:
                        loc:
                          type: array
                          items:
                            type: string
                        msg:
                          type: string
                        type:
                          type: string
  /api/extract-job-description:
    post:
      summary: Extract job description
      description: Extracts and processes job description from a provided URL
      operationId: extractJobDescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jobDescriptionLink
              properties:
                jobDescriptionLink:
                  type: string
                  format: uri
                  example: "https://example.com/job-posting"
      responses:
        '200':
          description: Successfully extracted job description
          content:
            application/json:
              schema:
                type: object
                description: Extracted and processed job description data
        '400':
          description: Bad request - invalid or missing job description URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid job description URL"
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Service is currently unavailable"
  /api/match-position:
    post:
      summary: Match position
      description: Analyzes a job description and returns matching information
      operationId: matchPosition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - profileId
                - jobDescription
              properties:
                profileId:
                  type: integer
                  description: ID of the user profile
                  example: 1
                jobDescription:
                  type: object
                  description: Job description data
              description: Job description data extracted from /extract-job-description
      responses:
        '200':
          description: Successfully matched position
          content:
            application/json:
              schema:
                type: object
                description: Job matching analysis result
        '400':
          description: Bad request - invalid job description data
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid job description data"
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Service is currently unavailable"
  /api/build-cv:
    post:
      summary: Build CV
      description: Generates a customized CV based on a job description
      operationId: buildCV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - profileId
                - jobDescription
              properties:
                profileId:
                  type: integer
                  description: ID of the user profile
                  example: 1
                jobDescription:
                  type: object
                  description: Job description data
              description: Job description data extracted from /extract-job-description
      responses:
        '200':
          description: Successfully built CV
          content:
            application/json:
              schema:
                type: object
                description: Generated CV content
        '400':
          description: Bad request - invalid job description data
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid job description data"
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Service is currently unavailable"
...
  /api/rewrite-achievement:
    post:
      summary: Rewrite achievement statement
      description: Rewrites and enhances an achievement statement to make it more impactful
      operationId: rewriteAchievement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - achievement_text
              properties:
                achievement_text:
                  type: string
                  description: The original achievement statement to rewrite
                  example: "I helped improve the team's productivity"
                style:
                  type: string
                  description: The style of rewriting
                  enum: ["professional", "concise", "impactful", "quantitative"]
                  default: "professional"
                  example: "professional"
                context:
                  type: string
                  description: Additional context about the achievement
                  example: "This was during my role as a project manager"
      responses:
        '200':
          description: Successfully rewritten achievement
          content:
            application/json:
              schema:
                type: object
                properties:
                  original_achievement:
                    type: string
                    example: "I helped improve the team's productivity"
                  rewritten_achievement:
                    type: string
                    example: "Enhanced team productivity by implementing streamlined workflows, resulting in a 25% increase in project delivery efficiency"
                  style:
                    type: string
                    example: "professional"
        '400':
          description: Bad request - empty or invalid achievement text
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Achievement text cannot be empty"
        '503':
          description: AI service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "An unexpected error occurred while rewriting the achievement statement."
  /api/rewrite-achievements:
    post:
      summary: Batch rewrite achievement statements
      description: Rewrites and enhances multiple achievement statements in one request
      operationId: rewriteAchievements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - achievements
              properties:
                achievements:
                  type: array
                  items:
                    type: string
                  example:
                    - "Improved system reliability by helping the team"
                    - "Worked on automation"
                style:
                  type: string
                  enum: ["professional", "concise", "impactful", "quantitative"]
                  default: "professional"
                context:
                  type: string
                  example: "During my role as a project manager"
      responses:
        '200':
          description: Successfully rewritten achievements
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        original_achievement:
                          type: string
                        rewritten_achievement:
                          type: string
                        style:
                          type: string
        '400':
          description: Bad request - empty or invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "At least one achievement must be provided"
  /api/experience:
    post:
      summary: Create experience entry
      description: Adds a new work experience entry to a user profile
      operationId: createExperience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperienceCreate'
      responses:
        '201':
          description: Experience created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperienceResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Profile with id 1 not found"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: array
                    items:
                      type: object
...
  /api/analyze-gaps:
    post:
      summary: Analyze experience gaps
      description: Identifies gaps between candidate's profile and job requirements with severity levels
      operationId: analyzeGaps
      parameters:
        - name: profile_id
          in: query
          required: true
          schema:
            type: integer
            example: 1
          description: ID of the user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobDescriptionResponse'
      responses:
        '200':
          description: Successfully analyzed gaps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GapAnalysisResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Profile with id 1 not found"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: array
                    items:
                      type: object
components:
  schemas:
    ProfileCreate:
      type: object
      required:
        - first_name
        - last_name
        - email
        - city
        - state
        - country
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 20
        city:
          type: string
          minLength: 1
          maxLength: 100
        state:
          type: string
          minLength: 1
          maxLength: 100
        country:
          type: string
          minLength: 1
          maxLength: 100
        summary:
          type: string
          nullable: true
    ProfileResponse:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        country:
          type: string
        summary:
          type: string
          nullable: true
...
    ExperienceCreate:
      type: object
      required:
        - profile_id
        - job_title
        - company
        - start_date
      properties:
        profile_id:
          type: integer
        job_title:
          type: string
          minLength: 1
          maxLength: 255
        company:
          type: string
          minLength: 1
          maxLength: 255
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
    ExperienceResponse:
      type: object
      properties:
        id:
          type: integer
        profile_id:
          type: integer
        job_title:
          type: string
        company:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
    Gap:
      type: object
      required:
        - gap_text
        - severity
        - suggestion
      properties:
        gap_text:
          type: string
          description: Description of the experience or responsibility gap
          example: "Missing evidence of leadership experience"
        severity:
          type: string
          description: Severity level of the gap
          enum:
            - Critical
            - Important
            - Nice-to-have
          example: "Important"
        suggestion:
          type: string
          description: Specific suggestion on how to address this gap
          example: "Add details about team leadership roles in your experience section"
    GapAnalysisResponse:
      type: object
      required:
        - gaps
      properties:
        gaps:
          type: array
          description: List of identified gaps between candidate and job requirements
          items:
            $ref: '#/components/schemas/Gap'